# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet App API",
        "description": "API documentation with Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List all transactions",
                "operationId": "eb94eb985d1e57150ce893ab651750cb",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "No transactions found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create a new transaction",
                "operationId": "cce56c116ea65120c279986f0865f206",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "value": {
                                        "description": "Transaction amount to be sent",
                                        "type": "number",
                                        "format": "float",
                                        "default": 100
                                    },
                                    "payer_id": {
                                        "description": "Id of who is sending the amount in this transaction",
                                        "type": "integer",
                                        "default": 4
                                    },
                                    "payee_id": {
                                        "description": "Id of who is receiving the value in this transaction",
                                        "type": "integer",
                                        "default": 15
                                    },
                                    "status": {
                                        "description": "Transaction status",
                                        "type": "string",
                                        "default": "in progress"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction by ID",
                "operationId": "f71625ec711512ca4bb2a8f96562b14b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Transaction not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    }
}